requirements.txt
django == 5.0.2
social-auth-app-django == 5.4.0
pillow == 10.2.0
schedule == 1.2.1
django-import-export == 3.3.7